+++ bankfeeds/RELEASES.md
@@ -679,3 +679,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-bankfeeds/0.33.1 - bankfeeds
+
+## 2023-08-17 00:15:25
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.73.1 (2.84.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] bankfeeds
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-bankfeeds/0.34.0 - bankfeeds